part_2-math/02_mini-projects/06-octahedron-opengl-pygame/requirements.txt
pygame==2.5.2
PyOpenGL==3.1.7
PyOpenGL-accelerate==3.1.7
vec2d==0.1.1
vec3d==0.1.0
